requirements.txt
flask
requests
elasticsearch==7.12.0
lycium~=0.2.15
PyYaml
pydantic
pycurl
dataclasses
dataclasses-json
asyncio
pycrypto
outcome
represent
pinyin
pymssql==2.2.5
demjson3
sqlalchemy
gmssl
asyncpg
pyjwt~=1.7.1
bcrypt
xmltodict
psycopg2-binary
tornado
cachetools
async-cached
aiohttp
aliyun-python-sdk-core-v3
pika==1.1.0
